Title: 앞에서부터 삭제하기 2, Time: 34ms, Memory: 1MB, Status: Wrong Answer - CodeTree

+++ 240403/앞에서부터 삭제하기 2/delete-it-from-the-beginning-2.cpp
@@ -22,7 +22,7 @@ int main() {
     }
 
     for(int i = 0; i < N - 2; i++){
-        int temp_sum;
+        double temp_sum;
         sum -= arr[i];
         if(-arr[i] == pq.top()){
             pq.pop();
